Move responsibility for locating plugin service registries from `core` to `build-state`.
